Implement modifiers
This is an important feature to support extensible weapons.
Every subclass of BaseWeapon can be modified by modifiers through
registering to the event of interest.

To implement a modifier, you need to inherit `IModifier<T>` where
T should be the type of weapon you want to modifier. To register
the modifier you just created, you need to create an instance of
it and pass it to the `RegisterModifier` function of the weapon.